string + number exception

+++ template-inheritance/Stationary.h
@@ -3,16 +3,16 @@ class Stationary : public VMachine
 {
 public:
 	Stationary();
-	Stationary(int powerSupply, string model);
+	Stationary(double powerSupply, string model);
 	Stationary(const Stationary& object);
 	~Stationary();
 
-	int getPowerSupply() const;
-	void setPowerSupply(int powerSupply);
+	double getPowerSupply() const;
+	void setPowerSupply(double powerSupply);
 
 	friend istream& operator >> (istream& in, Stationary& obj);
 	friend ostream& operator << (ostream& os, Stationary& obj);
 
 private:
-	int powerSupply;
+	double powerSupply;
 };
